[PCX-429] Move package name mode to input
