+++ payment/src/main/java/net/optile/payment/ui/widget/DateWidget.java
@@ -56,7 +56,6 @@ public final class DateWidget extends InputLayoutWidget implements ThemedDialogL
                 showDateDialogFragment();
             }
         });
-        setReducedView();
     }
 
     /**
