+++ payment/src/main/java/net/optile/payment/ui/widget/input/TextInputMode.java
@@ -6,7 +6,7 @@
  * See the LICENSE file for more information.
  */
 
-package net.optile.payment.ui.widget.mode;
+package net.optile.payment.ui.widget.input;
 
 import com.google.android.material.textfield.TextInputEditText;
