+++ payment/src/main/res/raw/validations.json
@@ -181,7 +181,7 @@
         "items": [
             {
                 "type": "bic",
-                "hide": false
+                "hide": true
             }
         ]
     }
